demo_pipeline/pyproject.toml
[project]
name = "demo-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.9.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
]
